+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.1] - 2021-05-02
 ### Added
+* Service contract definition [#3](https://github.com/cf-training-springboot-2019/spaceship-manager/issues/3)
+* Domain entity model [#2](https://github.com/cf-training-springboot-2019/spaceship-manager/issues/2)
 * Spaceover project standardized documentation [#1](https://github.com/cf-training-springboot-2019/spaceship-manager/issues/1)
 ### Changed
 ### Deprecated